+++ src/Doctrine/IdType.php
@@ -4,52 +4,16 @@ declare(strict_types=1);
 
 namespace DigitalCraftsman\Ids\Doctrine;
 
-use DigitalCraftsman\Ids\ValueObject\Id;
+use DigitalCraftsman\SelfAwareNormalizers\Doctrine\StringNormalizableType;
 use Doctrine\DBAL\Platforms\AbstractPlatform;
-use Doctrine\DBAL\Types\Type;
 
-abstract class IdType extends Type
+abstract class IdType extends StringNormalizableType
 {
-    abstract public static function getTypeName(): string;
-
-    /** @return class-string<Id> */
-    abstract public static function getClass(): string;
-
-    /** @codeCoverageIgnore */
+    /**
+     * @codeCoverageIgnore
+     */
     public function getSQLDeclaration(array $column, AbstractPlatform $platform): string
     {
         return $platform->getGuidTypeDeclarationSQL($column);
-    }
-
-    /** @param ?Id $value */
-    public function convertToDatabaseValue($value, AbstractPlatform $platform): ?string
-    {
-        if ($value === null) {
-            return null;
-        }
-
-        return (string) $value;
-    }
-
-    /** @param ?string $value */
-    public function convertToPHPValue($value, AbstractPlatform $platform): ?Id
-    {
-        if ($value === null) {
-            return null;
-        }
-
-        return static::getClass()::fromString($value);
-    }
-
-    /** @codeCoverageIgnore */
-    public function getName(): string
-    {
-        return static::getTypeName();
-    }
-
-    /** @codeCoverageIgnore */
-    public function requiresSQLCommentHint(AbstractPlatform $platform): bool
-    {
-        return true;
     }
 }