+++ tests/Test/Doctrine/UserIdListType.php
@@ -4,11 +4,10 @@ declare(strict_types=1);
 
 namespace DigitalCraftsman\Ids\Test\Doctrine;
 
-use DigitalCraftsman\Ids\Doctrine\IdListType;
-use DigitalCraftsman\Ids\Test\ValueObject\UserId;
 use DigitalCraftsman\Ids\Test\ValueObject\UserIdList;
+use DigitalCraftsman\SelfAwareNormalizers\Doctrine\ArrayNormalizableType;
 
-final class UserIdListType extends IdListType
+final class UserIdListType extends ArrayNormalizableType
 {
     public static function getTypeName(): string
     {
@@ -18,10 +17,5 @@ final class UserIdListType extends IdListType
     public static function getClass(): string
     {
         return UserIdList::class;
-    }
-
-    public static function getIdClass(): string
-    {
-        return UserId::class;
     }
 }